+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Preview release
 
 - Removed auto-scroll as it did not work as expected on a smartphone.
 - Renamed Button "New list" to "Clear list".
+- Added and changed some texts.
 
 ## 0.9.0 (2022-04-14)